scanline filling almost work
It works with concave and self-intersecting polygons. However, with
enough segments, it crash for some reasons. The case of horizontal
segments may not be handled right now and may produce unexpected
results.

The Bresenham-like algorithm used for following the edges is not exactly
the same as the one in segment_rasterize, hence a thin black like on the
edges.

It could also be optimized by sorting the active edge list with an
algorithm optimized for almost-already-sorted lists.

Signed-off-by: Celelibi <celelibi@gmail.com>